refactor(compiler): organize artifact graph modules
