+++ packages/compiler/package.json
@@ -48,6 +48,7 @@
     }
   },
   "dependencies": {
+    "@videojs/utils": "workspace:*",
     "kleur": "^4.1.5",
     "typescript": "6.0.2"
   },
