+++ packages/compiler/src/artifacts/index.ts
@@ -1,0 +1,23 @@
+export {
+  type BuildArtifactGraphOptions,
+  type BuildArtifactGraphResult,
+  buildArtifactGraph,
+} from './build';
+export {
+  type ArtifactDefinition,
+  type ArtifactMetadata,
+  type ArtifactMetadataValue,
+  type ArtifactResources,
+  type ArtifactSymbols,
+  defineArtifact,
+} from './definition';
+export {
+  ARTIFACT_GRAPH_VERSION,
+  type ArtifactClosure,
+  type ArtifactDependencies,
+  type ArtifactFile,
+  type ArtifactGraph,
+  type ArtifactGraphNode,
+  resolveArtifactClosure,
+  serializeArtifactGraph,
+} from './graph';
