+++ packages/compiler/tsdown.config.ts
@@ -6,7 +6,7 @@ export default defineConfig({
   entry: {
     index: './src/index.ts',
     ast: './src/ast.ts',
-    artifacts: './src/artifacts.ts',
+    artifacts: './src/artifacts/index.ts',
     'bundlers/vite': './src/bundlers/vite.ts',
   },
   platform: 'node',
